fix sidebar staff
