+++ src/app/(staff)/staff/StaffLayout.jsx
@@ -3,6 +3,7 @@
 import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 import Sidebar from '../../../components/Sidebar';
+import sidebarInfo from './sidebarInfo';
 
 const StaffLayout = () => {
     const router = useRouter();
@@ -29,40 +30,40 @@ const StaffLayout = () => {
         }
     ];
 
-    const sidebarItems = [
-        {
-            id: 'blog-management',
-            label: 'Blog Management',
-            icon: <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 20H5a2 2 0 01-2-2V6a2 2 0 012-2h10a2 2 0 012 2v1m2 13a2 2 0 01-2-2V7m2 13a2 2 0 002-2V9a2 2 0 00-2-2h-2m-4-3H9M7 16h6M7 8h6v4H7V8z" />
-            </svg>,
-            route: '/authen/staff'
-        },
-        {
-            id: 'donation-events',
-            label: 'Donation Events',
-            icon: <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z" />
-            </svg>,
-            route: '/authen/staff/event'
-        },
-        {
-            id: 'emergency-requests',
-            label: 'Emergency Requests',
-            icon: <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z" />
-            </svg>,
-            route: '/authen/staff/emergency'
-        },
-        {
-            id: 'blood-stock',
-            label: 'Blood Stock',
-            icon: <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 11H5m14 0a2 2 0 012 2v6a2 2 0 01-2 2H5a2 2 0 01-2-2v-6a2 2 0 012-2m14 0V9a2 2 0 00-2-2M5 11V9a2 2 0 012-2m0 0V5a2 2 0 012-2h6a2 2 0 012 2v2M7 7h10" />
-            </svg>,
-            route: '/authen/staff/stock'
-        }
-    ];
+    // const sidebarItems = [
+    //     {
+    //         id: 'blog-management',
+    //         label: 'Blog Management',
+    //         icon: <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+    //             <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 20H5a2 2 0 01-2-2V6a2 2 0 012-2h10a2 2 0 012 2v1m2 13a2 2 0 01-2-2V7m2 13a2 2 0 002-2V9a2 2 0 00-2-2h-2m-4-3H9M7 16h6M7 8h6v4H7V8z" />
+    //         </svg>,
+    //         route: '../staff'
+    //     },
+    //     {
+    //         id: 'donation-events',
+    //         label: 'Donation Events',
+    //         icon: <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+    //             <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z" />
+    //         </svg>,
+    //         route: '../staff/event'
+    //     },
+    //     {
+    //         id: 'emergency-requests',
+    //         label: 'Emergency Requests',
+    //         icon: <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+    //             <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z" />
+    //         </svg>,
+    //         route: '../staff/emergency'
+    //     },
+    //     {
+    //         id: 'blood-stock',
+    //         label: 'Blood Stock',
+    //         icon: <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+    //             <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 11H5m14 0a2 2 0 012 2v6a2 2 0 01-2 2H5a2 2 0 01-2-2v-6a2 2 0 012-2m14 0V9a2 2 0 00-2-2M5 11V9a2 2 0 012-2m0 0V5a2 2 0 012-2h6a2 2 0 012 2v2M7 7h10" />
+    //         </svg>,
+    //         route: '../staff/stock'
+    //     }
+    // ];
 
     // User info for staff
     const userInfo = {
@@ -84,7 +85,7 @@ const StaffLayout = () => {
     return (
         <div className="flex h-screen bg-gray-50 border border-gray-200">
             {/* Sidebar */}
-            <Sidebar items={sidebarItems} userInfo={userInfo} portalType="Staff" initialTab={activeTab} />
+            <Sidebar items={sidebarInfo()} userInfo={userInfo} portalType="Staff" initialTab={activeTab} />
 
             {/* Main Content */}
             <div className="flex-1 flex flex-col">
