+++ src/app/(staff)/staff/userInfo.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+function userInfo() {
+    const userInfo = {
+        name: 'Nguyen Thi B',
+        role: 'Staff Member',
+        initials: 'NT'
+    };
+    return (
+        userInfo
+    )
+}
+
+export default userInfo
